+++ aula4/cliente.json
@@ -1,0 +1,11 @@
+{
+  "nome": "Joao",
+  "email": "joao@firma.com",
+  "telefone": ["11223344", "11922334453"],
+  "endereco": {
+    "rua": "R. Joseph Climber",
+    "numero": 404,
+    "apartamento": true,
+    "complemento": "ap 934"
+  }
+}
