+++ aula4/leitura.js
@@ -1,0 +1,15 @@
+const dados = require("./cliente.json");
+
+console.log(dados);
+console.log(typeof dados);
+
+const clienteEmString = JSON.stringify(dados);
+
+console.log(clienteEmString);
+console.log(typeof clienteEmString);
+
+console.log(clienteEmString.nome);
+
+const objetoCliente = JSON.parse(clienteEmString);
+
+console.log(objetoCliente);
